requirements.txt
# Phase 1
python-dotenv==1.0.0

# Phase 2
python-telegram-bot==20.7
